Cuke showing RestfulRails paginating correctly
This fixes a bug where RestfulRails would blow up when
attempting to paginate, discovered via failing cucumber feature.
The lib and specs are modified accordingly.

This commit also removes the artificial pagination default of 15,
instead relying on will_paginate's default of 30. This change
should be a separate commit, but is not.